Add common add Dialog component
